backend/requirements.txt
fastapi
sqlalchemy
uvicorn
fish-audio-sdk
authlib
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]
python-dotenv
deepgram-sdk
google-genai